chunker/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-docx>=0.8.11
PyPDF2>=3.0.1
python-multipart>=0.0.6
chonkie>=1.1.0
sentence-transformers>=2.2.0
pymilvus>=2.5.0
tqdm>=4.64.0
torch>=2.0.0
transformers>=4.30.0
extra-streamlit-components>=0.1.70